Move files to match the renamed classes
