+++ src/analyzer.ts
@@ -19,23 +19,23 @@ import * as path from 'path';
 import {Document, InlineParsedDocument, LocationOffset, ScannedDocument, ScannedElement, ScannedFeature, ScannedImport, correctSourceRange} from './ast/ast';
 import {CssParser} from './css/css-parser';
 import {Severity, Warning, WarningCarryingException} from './editor-service';
-import {Scanner} from './entity/entity-finder';
-import {scan} from './entity/find-entities';
-import {HtmlImportScanner} from './html/html-import-finder';
+import {HtmlImportScanner} from './html/html-import-scanner';
 import {HtmlParser} from './html/html-parser';
-import {HtmlScriptScanner} from './html/html-script-finder';
-import {HtmlStyleScanner} from './html/html-style-finder';
+import {HtmlScriptScanner} from './html/html-script-scanner';
+import {HtmlStyleScanner} from './html/html-style-scanner';
 import {JavaScriptParser} from './javascript/javascript-parser';
 import {JsonParser} from './json/json-parser';
 import {ParsedDocument} from './parser/document';
 import {Parser} from './parser/parser';
 import {Measurement, TelemetryTracker} from './perf/telemetry';
-import {BehaviorScanner} from './polymer/behavior-finder';
-import {DomModuleScanner} from './polymer/dom-module-finder';
-import {PolymerElementScanner} from './polymer/polymer-element-finder';
+import {BehaviorScanner} from './polymer/behavior-scanner';
+import {DomModuleScanner} from './polymer/dom-module-scanner';
+import {PolymerElementScanner} from './polymer/polymer-element-scanner';
+import {scan} from './scanning/scan';
+import {Scanner} from './scanning/scanner';
 import {UrlLoader} from './url-loader/url-loader';
 import {UrlResolver} from './url-loader/url-resolver';
-import {ElementScanner as VanillaElementScanner} from './vanilla-custom-elements/element-finder';
+import {ElementScanner as VanillaElementScanner} from './vanilla-custom-elements/element-scanner';
 
 export interface Options {
   urlLoader: UrlLoader;
