+++ src/polymer/feature-scanner.ts
