+++ src/polymer/polymer-element-scanner.ts
@@ -19,7 +19,7 @@ import * as astValue from '../javascript/ast-value';
 import {Visitor} from '../javascript/estree-visitor';
 import * as esutil from '../javascript/esutil';
 import {JavaScriptDocument} from '../javascript/javascript-document';
-import {JavaScriptScanner} from '../javascript/javascript-entity-finder';
+import {JavaScriptScanner} from '../javascript/javascript-scanner';
 
 import {PropertyHandlers, declarationPropertyHandlers} from './declaration-property-handlers';
 import * as docs from './docs';
