+++ src/test/polymer/behavior-scanner_test.ts
@@ -22,7 +22,7 @@ import {Visitor} from '../../javascript/estree-visitor';
 import {JavaScriptDocument} from '../../javascript/javascript-document';
 import {JavaScriptParser} from '../../javascript/javascript-parser';
 import {ScannedBehavior} from '../../polymer/behavior-descriptor';
-import {BehaviorScanner} from '../../polymer/behavior-finder';
+import {BehaviorScanner} from '../../polymer/behavior-scanner';
 
 suite('BehaviorScanner', () => {
