+++ src/vanilla-custom-elements/element-scanner.ts
@@ -19,7 +19,7 @@ import * as astValue from '../javascript/ast-value';
 import {Visitor} from '../javascript/estree-visitor';
 import * as esutil from '../javascript/esutil';
 import {JavaScriptDocument} from '../javascript/javascript-document';
-import {JavaScriptScanner} from '../javascript/javascript-entity-finder';
+import {JavaScriptScanner} from '../javascript/javascript-scanner';
 import * as jsdoc from '../javascript/jsdoc';
 
 export interface ScannedAttribute extends ScannedFeature {
